TypeAutoFields is now JSONSchemaForm
