+++ frontend/json_form/fields/Secret.js
@@ -1,0 +1,31 @@
+import React from "react";
+import { Input as ChakraInput, Flex, FormLabel } from "@chakra-ui/react";
+import { useEditorColorMode } from "chains/editor/useColorMode";
+import { getLabel } from "json_form/utils";
+
+export const Secret = ({ name, field, isRequired, value, onChange }) => {
+  const colorMode = useEditorColorMode();
+  const handleChange = React.useCallback(
+    (event) => {
+      onChange(name, event.target.value);
+    },
+    [field, onChange]
+  );
+
+  return (
+    <Flex justifyContent="space-between" wrap="wrap" width="100%">
+      <FormLabel justify="start" whiteSpace="nowrap">
+        {getLabel(name)}
+      </FormLabel>
+      <ChakraInput
+        value={value || ""}
+        onChange={handleChange}
+        px={1}
+        py={2}
+        type={"password"}
+        sx={field.style || { width: 200 }}
+        {...colorMode.input}
+      />
+    </Flex>
+  );
+};
